packages/agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.7",
    "langgraph>=1.0.3",
    "langchain-openai>=0.3.14",
    "langchain-core>=1.0.5",
    "python-dotenv>=1.0.0",
    "mcp>=1.21.1",
]
